+++ CHANGELOG.md
@@ -1,3 +1,10 @@
+# You're in Grave Danger 2.0.0-beta.9
+
+### Changes
+* Added `/yigd whitelist list` command to list all whitelisted/blacklisted players
+
+---
+
 # You're in Grave Danger 2.0.0-beta.8
 
 ### Changes